Refactored richtext component into a server side and client side components to optimize performance
